REF: Move load_graphs() to a new separate module cmtklib.bids.network
